feat: Add admin  and Instructor controllers and routes and refatoring the index file
This commit adds a new controller and routes for managing admin profiles. It includes functions for creating an admin profile, getting the admin profile, editing the admin profile, and deleting the admin profile.

This commit adds a new controller and routes for managing instructor profiles. It includes functions for creating an instructor profile, getting the instructor profile, editing the instructor profile, and deleting the instrutor profile.

Remove mongoose and chnaged routes from the index file